pyproject.toml
[project]
name = "aethernet-core"
version = "1.0.0"
description = "Part of Slet Infrastructure"
authors = [
    { name = "Walter Kerrigan", email = "esolment@gmail.com" },
]
readme = "README.md"
requires-python = ">=3.12, <3.14"
license = { file = "LICENSE" }
dependencies = [
    "basex-python==0.1.0",
    "pycryptodome>=3.23.0,<4",
    "msgpack>=1.1.2,<2",
    "httpx>=0.28.1,<0.29",
    "websockets>=16.0,<16.1",
    "twine>=6.2.0",
    "build>=1.4.3",
]

[project.optional-dependencies]
tests = [
    "pytest>=9.0,<10",
    "pytest-mock>=3.15,<4",
    "pytest-asyncio>=1.3,<2",
    "pytest-cov>=7.1,<8",
    "hypothesis>=6.152,<7",
]

lint = [
    "ruff>=0.15,<1",
    "black>=26.3,<27",
    "mypy>=1.20,<2"
]

dev = [
    "aethernet-core[tests,lint]",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "-ra",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-fail-under=50"
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.mypy]
exclude = ["tests/"]
ignore_missing_imports = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aethernet"]
